+++ MiniApp/Classes/ui/MiniAppViewController.swift
@@ -231,11 +231,7 @@ extension MiniAppViewController: MiniAppNavigationDelegate {
 
     /// will be used soon when knowing the cases to react to
     public func miniAppNavigation(shouldOpen url: URL, with responseHandler: @escaping MiniAppNavigationResponseHandler) {
-//        MiniAppExternalWebViewController.presentModally(url: url, externalLinkResponseHandler: { url in
-//            if url.absoluteString == "miniapp://close" {
-//                // dismiss
-//            }
-//        })
+        // Implement navigation handling when necessary (todo raises a swiftlint error)
     }
 
 }